[syzkaller][fidl] Fix Syzkaller IR to keep in-sync with C++.
Syzkaller's fidlgen was producing defines such as
`CodecBufferDataTagvmo`, but the corresponding fidl.h files produce
defines of the form `CodecBufferDataTag_vmo`, which caused Syzkaller to
be unable to update its own headers appropriately.

Long-term I'd like to sit down and think about a better way to prevent
drift here, since Syzkaller descriptions are
kind-of-their-own-thing-but-kind-of-not; however, this should unbreak
Syzkaller for now.

Test: manually ran FIDLgen

Change-Id: Ia7d29504586f6eea36585d9e1d1aaf2ca4827b48