[escher] Another round of pre-shadow-volumes changes.
The biggest changes are to RenderQueue.  RenderQueueItem now contains
multiple RenderQueueItem::Funcs which are dynamically selected during
traversal.  The selection is based on info in the newly-introduced
RenderQueueContext, which is also passed to each func invocation.

(it's not clear how useful having multiple funcs-per-item will turn
out to be, but the ability to pass a context into each function is
crucial).

TEST=escher_unittests (added test for multiple render funcs)

Change-Id: I8cb5b92871d6d4bf2a4a8b9e68a87f69b9225037