[ledger] Introduce infrastructure to remove systematic return of Status.
This introduces a new base interface for Ledger user-facing FIDL
services: ledger.ErrorNotifier

Interfaces that implements ErrorNotifer will not return Status for
exceptional errors. Instead, they will close the connection with an
epitaph containing a ledger::Status.

To allow this API change, this CL introduces code generation to generate
a Proxy and a Delegate for every Ledger interface that extends
ledger.ErrorNotifier. It allows to implement the Delegate that still
return a Status, and have the Proxy close the connection with the right
error. The Delegate can be associated with multiple proxy which will
allow to still use a single instance of the implementation but multiple
connections with scoped errors.

LE-618

TEST=ErrorNotifierTest.*, existing tests still pass.

Change-Id: I06dd92e46c626816f3fb1433f0fbd2ae8948680b