[fidl] Allow reserved keywords as identifiers
Instead of representing special identifiers such as 'handle', or
'request' as a specific kind, we represent them as identifiers with a
subkind. This makes it possible to be more precise on how these tokens
are used in parsing, and specifically makes it possible to allow these
to be used where we expect grammar identifiers.

Test: make -j 12 tools HOST_USE_ASAN=true && ./build-x64/host_tests/fidl-compiler-test
Change-Id: I816faee504d143c20091c6e60d0a2d0ce4ae6216