[netcfg] Generate stable network interface names
Implemented stable interface naming methodology in netcfg, which will
generate the names from either the topological path for onboard devices
or the mac address for usb devices. The configuration are stored in a
config file which is persistent across the bootup of the OS. Only if
the added interfaces has no match in the configuration file, a new name
will be generated and stored in the configuration file.

Test: - fx run-test netcfg_bin_test_rustc
  - fx build;
  - fx shell ifconfig
    (snip)
    lo	HWaddr  Id:1
    ethxd8	HWaddr 8c:ae:4c:ee:54:d8 Id:2
    wlanxf9	HWaddr 70:f1:1c:5:2e:f9 Id:3

  - fx reboot; fx shell ifconfig
    (snip)
    lo	HWaddr  Id:1
    ethxd8	HWaddr 8c:ae:4c:ee:54:d8 Id:2
    wlanxf9	HWaddr 70:f1:1c:5:2e:f9 Id:3
  - # unplug dongle; plug back in
    - fx shell ifconfig
    (snip)
    lo	HWaddr  Id:1
    ethxd8	HWaddr 8c:ae:4c:ee:54:d8 Id:2
    wlanxf9	HWaddr 70:f1:1c:5:2e:f9 Id:4

Change-Id: Ia673ba557cc3e180e8c3a343f831148443bd0e89