[zircon][usb-request] Do not use the usb-request ops from protocol
The usb protocol got too big with the usb-request ops. The
usb-request ops are more about managing a vmo and have little to do
with the usb protocol as such. This changeset removes the usage of
usb-request protocol ops and uses usb-request static library instead.
The usb ops will be removed in a later changeset once all the usage is
removed.

Test: lsusb on vim2. Tested a usb mass storage device read using dd.

Change-Id: I9325eb16cb5dc3df64941cb83df0f1dfae9ba70f