[mediaplayer] change over to new allocation scheme
This CL changes the allocation scheme of mediaplayer. Previously,
allocators were only exposed in terms of malloc/free-style methods,
and allocators were sometimes exposed by outputs to be used by
inputs. The new scheme has the nodes providing configuration info
for each input and output. For a given connection, the configuration
data for the output and input are considered by PayloadManager to
produce an allocation strategy for the connection, which
PayloadManager then implements.

FidlDecoder is neglected for the moment and is broken by this CL.
The changes required there are extensive and have been left to a
later CL. FidlDecoder isn't currently in use anyway, so we don't
lose anything by breaking it temporarily.

The decoder code that integrate mediacodec is turned off for now.
Previously, this was done by disabling a type conversion. This CL
just refrains from adding the relevant decoder factory from a list.

TEST: No behavior change. Manual and CQ regression tests.
Change-Id: Id9e130ab35a6abb57724ca7eb9b0f66850af672e