[debugger] Enhance symbol resolution.
Moves InputLocation to the symbol library and adds a function to resolve
it to Location objects on ModuleSymbols, ProcessSymbols, and
TargetSymbols.

In a later phase this will replace several of the existing functions
that look up symbol information. Currently this exists in addition to
make this patch smaller. The tests for those other functions will be
moved to test the new function at that time.

The breakpoint and list code now use the new code path which reduces
some duplicated code to resolve different types of inputs.

The purpose of this change is to allow this same type of input
resolution to be used for other commands, for example "x <symbol>" or
"disassemble main". It will be added to other commands in a future
pass.

TEST=unit,manual

Change-Id: I09de54cb4716cbe64767c3e672ca6c56c67a010e