[blobfs] [lz4] Use dest buffer max size in compression calls
Currently host-created blobs are not being compressed
due to a failure on Compressor::Update where the dest
buffer max size is not large enough. This is because the
remaining size of the dest buffer is being passed, rather
than the max capacity. This change passes in the max
capacity instead.

This bug does not affect target-created blobs to the same
extent, since right now compression is broken up into
chunks of 8kb maximum.

ZX-2713 #comment In progress

Test: TestCompressorBufferTooSmall in blobfs-test

Change-Id: Ib6c6ccf5bb1ac05c1312b0627a7acc97c44b434b