[wlan][brcmfmac] Add wlanif disconnect support
Add support for DEAUTHENTICATE.request/confirm,
DISASSOCIATE.request/confirm, and DISASSOCIATE.indication.

On disconnect (DEAUTH/DISASSOC) requests, a response is delayed until
either a timeout occurs or confirmation of the disassociation status
is received from the firmware. This helps to avoid confusion during
event processing when a connect is initiated immediately following a
disconnect.

Includes a workaround for NET-1574

NET-984

Test: Associated/Disassociated 20x and verified signficantly
reduced failure rates.

Change-Id: Ia6a7ea1141426dfec2a96b6164eeb531fc925f47