[garnet][usb-request] Do not use the usb-request ops from protocol
The usb protocol got too big with the usb-request ops. The
usb-request ops are more about managing a vmo and have little to do
with the usb protocol as such. This changeset removes the usage of
usb-request protocol ops and uses usb-request static library instead.
The usb ops will be removed in a later changeset once all the usage is
removed.

Test: Build

Change-Id: I21a754dddae0e9232605682d998b6b10b2012a34