[zircon][usb-function] Use usb-request functions directly
The usb-request ops are more about managing a vmo and have little to do
with the usb function protocol as such. This changeset removes the usage of
usb-request protocol ops and uses usb-request static library instead.
The ops from usb-function protocol  will be removed in a later changeset
once all the usage is removed.

Test:Build

Change-Id: Ib95392e2b9f7f6a91f087943af02591348cd7ca9