[ddk] Start moving platform-defs.h out of the "protocols" directory
This file does not actually contain a protocol, and the "protocols"
directly will likely go away when we auto generate the protocols.

This CL copies platform-defs.h to system/ulib/ddk/include/ddk/
The old copy will be removed after the garnet drivers are updated to use
the new location.

TEST: zircon builds.
Change-Id: Iaa17f2ff0164f07ce9b2d733ad6ec001c6fdfcb8