[optee] Migrate GetInfo IOCTL to FIDL
The GetInfo method provides a TEE OS identifier, the TEE OS revision and
indication whether the API is Global Platform Compliant. If it is Global
Platform Compliant, it can be used in conjunction with the TEE Client
API. This change deprecates the GetDescription IOCTL interface and
migrates the TEE Client API to use the FIDL interface. The rename from
GetDescription to GetInfo was to align with other drivers.

Test: Ran tee-test on Astro and ensured that context could still be
opened.

Change-Id: I0d5e51face217c17672d3dd0d4d903afd8a9b92f