[debugger] Supprt printing unions.
Generalized the old "StructClass" to a "Collection" type that also
supports unions. Most of the change is just a rename for this.

The first place with logic change in in the DWARF symbol factory where
it now creates this Collection class for unions (otherwise the code is
unchanged alghough it moved).

The second place is an added test for union formatting. The actual
formatting code did not need to change since the union uses the same
underlying constructs as structs and classes.

TEST=unit,manual
DX-509

Change-Id: Idbefdb50626b0851beda1ddbaca22c6bf6f74eb6