[ota] Add OTA test suite
This adds an integration test for Over the Air updates. At a high
level, it extends the current system image by inserting in a new
unique file into it. The test then is to trigger an OTA, and verify
the file exists in the latest system. See the README.md for more
details.

PKG-248 #done

TEST: Manually confirmed test pass when run with the configuration
described in the README.md.

Change-Id: I87b9d45885ba25503122d07a05ead1938ef8174c