[bt-gap] Add bonding serialization and unit tests
Added the initial bits of bonding data (de)serialization code to
the bt-gap component. The code is largely based on existing code in
bt-mgr with slight readability improvements. This is the first step
towards using the "stash" for bonding data storage.

This change also introduces a unit test binary for the bt-gap component.
The initial set of unit tests exercise the JSON serialization code
introduced in this change.

NET-1605 #progress
TEST=/pkgfs/packages/bt-gap-tests/0/bt-gap-unittests

Change-Id: I9c1365ab67eb17a393ccddfde67b00237c4af0bd