[musl][ldso] Support for XRay
XRay needs to retrieve the innermost VMAR to allow changing its
permissions when patching function entry and exit points. This is
currently not possible since the dynamic linker always closes the VMAR
handle. To support XRay, when the dynamic linker is being built for
XRay, we don't close the VMAR handle and provide a new runtime function
__sanitizer_change_code_protection which allows caller to change the
protection of its own code segment.

Test: manual
Change-Id: If6a780d29cce389068574cc8b50d0f2387f8e3a3