[crash] Pass exception port in crash analyzer interface
zx_task_resume() is deprecated. In order to use
zx_task_resume_from_exception(), the exception port which serviced the
exception must be plumbed through.

To accomplish this, pass it from crashsvc back over the channel to
devmgr, and have devmgr dispatch the port via fidl to the analyzer.

Additionally, fix a few places that weren't updated when
zx_channel_write() started always-consuming handles, and a couple of
ASSERTs in tests.

ZX-2720 #comment

Test: /system/test/sys/fidl-simple-test
Change-Id: I969676a67ee5321ea027e4232a9f1093a56da653