[kernel][socket] Enhancement to threshold based signalling on sockets.
If we bump up the socket Tx and Rx thresholds, de-assert the READ/WRITE
threshold signals if the amount of data queued for read or the amount
of space available to write falls below the new threshold.

ZX-2200.

Test: Added test cases for the above to the unit test to core/socket-test
to make sure the signals are being de-asserted as we bump up thresholds.

Change-Id: I5f1db55604f664afe67f2b423e13db85fe32b387