[debugger] take general locations for "x" and "di"
Converts "mem-read" ("x") and "disassemble" commands to accept a general
location rather than just a raw address. This allows you to say
  disassemble main
for example.

Both of these functions now default to the size of the input object if
it has a known size. This way "disassemble main" will output all
disassembly for the main function unless you override the default. The
mem-read command works the same way.

Adds a bunch of machinery to the input location resolver to get locations.

Enhanced the range handling of the CodeBlock. Enhances and fixes a bug
in command_utils.cc by taking a SymbolContext in the range comparison.

TEST=manual,unit

Change-Id: I857d3c8e866f98120508cd073594f01559c21a0f