[build] Fix incremental build cases with implicit stripping
When finalize_manifests.py writes stripped files (e.g. for toolchain
libraries), GN/Ninja doesn't know it changed that output file.  This
means an incremental build can re-run finalize_manifests.py but then
fail to re-run packaging and image-packing steps even though it
touched a stripped file and that causes the *next* Ninja run to see
that file from the depfile and run the packaging steps then.

This change fixes both the case when nothing really changed and the
case when something actually did change, re-running the steps that
need to be done and not the ones that don't.

Firstly, finalize_manifests.py now checks whether a stripped file
it's about to write would be identical to the file already there and
doesn't touch the file if it won't change.  Secondly, whenever it
actually did touch any stripped file, it forcibly updates the output
manifest files so that GN/Ninja will definitely re-run all the
packaging steps to pick up the changed files listed therein.

Bug: BLD-289 #done
Test: manual
Change-Id: Id1bc5af72906b12918215550822ef13fe116c5ac