[package system] large scale refactor & bugfix
This is an oversized change. It is hard to split down because while some
changes could be split out independently, many of the changes have
interdependent semantics changes that ultimately produce bug fixes. After an
evaluation of the level of effort and intermediate stability of chunking this
up, it is being proposed as-is.

- Introduce atonce, a package much like "singleflight" but simpler, and use
it to single-flight package updates, blob fetches, and system update checks.
- amber-srv and pm serve have their output adjusted to include status codes
(useful for diagnosis of ongoign networking issues and other side effect
behaviors). The truncation is also removed, as it simply gets in the way of
debugging.
- ambers unused startup delay is removed
- system update monitor now calls amber daemon methods directly, rather than
using a fidl client to the same pid as the fidl server.
- the fidl server now dispatches with some concurrency (numcpu)
- the syslog log writer now reports correct write lenghts to it's caller
- the syslog logs no longer include two timestamps
- the file logger is removed - superseeded by syslog features
- various dead code eliminations
- ipcserver is renamed control_server
- APIs use the same notion of name and version as pkgfs and packages
- update requests that identify a merkle version now install directly from
blobs, improving both performance and reliability, removing various edge
cases such as newer packages being present at the same TUF targets.
- amber-control now explicitly closes it's handle before exiting
- ipcclient is removed - it was only used by amber-control
- control server has many changed semantics:
  - most long running calls are asynchronous
	- some synchrnous APIs are removed
- a daemon-global activation watcher is introduced to centralize callback
processing for installed packages, and is used to solve signalling back to
waiting clients.
- the blob fetcher is removed, its functionality is now split between the
daemon and the sources that the blob repos came from.
- the daemons now unused source monitor facilities are removed.
- most concurrent state management is removed from the daemon, see atonce for
the most critical de-duplication work.
- the source interface is removed, source is now tufsource - this simplifies
further refactoring work.
- the update fetching process is split into two phases, first merkleroot
identification, followed by installation. this change facilitated the use of
the atonce abstraction, removing races between installation and observer
setup.
- package set is removed, as the abstraction was not being used.
- publishing a package now publishes the meta.far as a blob as well as a tuf
target.
- pkgfs now serves the package meta.far merkleroot as a file at /pkg/meta.
/pkg/meta has directory/file duality, but presents as a directory under
O_PATH (for stat, readdir, etc).
- system update no longer makes install requests for packages that are
already installed
- the system update monitor no longer allows parallel execution of an update
check.
- the system update no longer attempts to resolve package variants/versions
in numerical order, instead it binds to 0 as all other subsystems do. any
future usage of variants may have otherwise broken due to not being numeric,
and the behavior lacked determinism in the face of present invariants.
- pkgfs now reconnects to amber when necessary, and logs errors sending
messages to amber (such errors are still irrecoverable - this architecture
must be changed)
- pkgfs choosing to update the dynamic vs static index is now correct.
- pkgfs needs directory no longer supports writing new packages into it.
- pkgfs no longer opens blobs in the install dirs when O_PATH'd
- pkgfs more reliably reports EEXIST:
  - package meta.far writes for actviated packages always EEXIST
  - package meta.far writes for incomplete packages do not EEXIST, even if
  the meta.far blob does already exist.
- pkgfs now logs package activations and EEXISTS, to aid user understanding
- pkgfs unimplemented /metadata directory is removed.
- pkgfs now correctly lists the system_image package under /pkgfs/packages
- system update handling of the amber async update protocol is simplified
- pkgfs GC no longer misses dynamically installed updates that attain to the
static index. the static index now tracks both dynamic updates and it's
originally loaded data.
- pkgfs GC avoids taking the root directory lock, was not intended, nor does
it operate as a global pkgfs lock.
- pkgfs GC now frees directories that it opens once it is done reading from
them.

Bug: PKG-10 #done
Bug: PKG-69 #done
Bug: PKG-89 #obsolete
Bug: PKG-142 #done
Bug: PKG-154 #done
Bug: PKG-236 #comment clarified installation semantics
Test: manually tested OTAs and package pushes numerous times against multiple device targets, both arm64 and x64, using both `fx` tools and direct invocations of amber_ctl. pmd_pkgfs_test and amber_daemon_test updated and run in CQ.
Change-Id: I28bde6654acaa25ec1308c45b08e3c7b76904392