[debugger] Allow register fetching to fail.
Previously no registers would be returned if fetching any component
failed.

But just because one category failed doesn't mean we can't reply with the
other categories. This patch removes the full failure in the case of
fetch failure and returns as much as we can get.

This has the short-term effect of fixing register fetching and printing
because the kernel currently reports failure when fetching debug
registers.

TEST=manual

Change-Id: I4adbf7d29b5e2474bb1d448e4f068f224a309b94