[debugger] Separate out build ID parsing.
Build I Dparsing used to be in the debug_agent. This separates it out
into the debug helper library so it can be shared with the client. The
read interface has been virtualized so it can read in-memory or files on
disk.

A unit test is added for the parsing.

TEST=unit

Change-Id: I1bca474553ea970d2e7e5fd53e1fd988370581b7