[bt][intel] Fix initialization failures
* Send HCI_Reset and wait up to 10 seconds for HCI Command Complete.
Ignore spurious HCI events while waiting for the initial reset.

* Ignore "Unknown Command" error during initial HCI_Reset on controllers
using the "secure send" method as this usually means that the hardware
is in BOOTLOADER mode.

* Sleep for 2 seconds after sending the "Vendor Reset" command to wait
for the controller to switch to FIRMWARE mode. Don't send this command
on controllers that don't support the "secure send" method.

* Wait on both event and ACL channels for HCI events on controllers that
use the secure send method.

NET-1656 #done
TEST=Watch logs for driver initialization on platforms using Intel AC 726x
and 926x. The btintel driver should succeed to bind both after a cold
boot and a soft reboot. Repeat process multiple times.

Change-Id: Iafa9b50ebebc9464d40f53a2ea38717658d11edc