[dev][bt][usb] Fix write to invalid channel handle
As writing to an invalid channel handle now causes a crash, this
CL removes the assumption that we'll fail with a ZX_ERR_BAD_HANDLE
when handling received ACL data packets.

Also added additional log messages and switched to using the zxlog
macro.

NET-1583 #done
TEST=Force the driver to unbind while ACL data frames are buffered.
This should not cause a crash.

Change-Id: I93a56d14c4202d07fa7cba78fb71c37823875171