[ddk][mmio-buffer] Add mmio_buffer_init_physical
This secondary init method provides a way to map a physical region for
mmio use given a resource capable of creating a physical vmo. It is meant
to be used in the board drivers, as most drivers will not have access to
such a resource, but will instead be given the a VMO they need to map.

Tested: Booted vim2 with updated board driver (and a future change).
Change-Id: If63efbb735094a4527bb465b895c5f8a79e40a10