[musl][futex] Remove __wake(); inline it into its call sites
__wake() was a trivial wrapper around _zx_futex_wake().  Inlining it
should make the code clearer by removing a level of indirection.

The wrapper just converted negative wake counts to INT_MAX.  We now
use UINT32_MAX in those cases.  (UINT32_MAX is the largest wake count,
because wake counts are of type uint32_t.)

The only non-constant wake count was in pthread_rwlock_unlock().  In
that case, cnt was always >0, because it's set to 'val & 0x7fffffff'.

Bug: none
Test: none apart from CQ tests
Change-Id: I8a81d4bd282d5ca4e77b56f9a8534bd9ab7a2ac1