[bt][l2cap] FakeChannel: update SendCallback type
For stylistic consistency, update the ByteBuffer type argument to the
type template referenced by the SendCallback type. Update callers
accordingly.

Note that the new type is not identical to the old type. However, the
new type still prevents modification to the data held by the referent,
absent any explicit downcast.

While there: update the return type of SignalingChannel::BuildPacket()
in the same manner.

Bug: NET-1534 #done
Test: bt-host-unittests (on eve)

Change-Id: Id07f5b4b1eb88c420ac4a8994d71c50a35fd9119