[dev][usb-bus] Add a timeout for usb_control.
When plugging in a usb hub, it may disconnect briefly after beginning
enumeration.

If the disconnect occurs during the usb_device_add call, the control
request to get_descriptor will not complete. Since the xhci device manager
thread is stuck waiting on the enumeration to finish, it will not get to the
disconnect device command and clean up those requests.

This looks to be one of the cases causing ZX-2510.

TEST= unplugging and replugging usb hub and pressing power button soon
after.

ZX-2510 #comment

Change-Id: Ibd134e682cf050e90220545993e85f1ff3f38a40