[bt][l2cap] PDU::Reader: simplify semantics of ReadNext()
When the next chunk of data resides wholly within the current
fragment, update our internal state _before_ invoking the ReadFunc. In
particular, this ensures that we update the FragmentList::iterator
before the ReadFunc has the opportunity to destroy the PDU (the latter
of which would invalidate the iterator).

This enables callers to discard the underlying storage within their
ReadFunc, instead of having to defer that work until after the
callback. This also makes the behavior for the within-a-fragment, and
cross-fragment-boundaries cases, uniform.

While there: use the new semantics to simplify some calling code in
SocketChannelRelay.

Bug: NET-1483
Test: bt-host-unittests (on eve)
Change-Id: I79055a3151531c1425ef3334c01e9c0bc46ea893