[zxdb] Stub out MinidumpRemoteAPI
We want to introduce support for debugging minidumps rather than live
systems. This lays the framework for that by providing an alternate
connection method for Session that doesn't connect to an agent, but
rather swaps in a new remote API which will operate directly on the dump
file. At the moment that API is actually a stub.

Test: Unit tests continue to pass
Change-Id: Iae2dbe3e4ae900738a56097c5c10df6f11b97623