[fidl] Redo defaults checking in const declarations
Previously, there were many cases which were not covered, as discussed
in FTP-003 on Defaults. For instance, it was possible to have

    const string s = 5;

Decomposing const declaration type checking into

- can the type be a const at all? e.g. arrays cannot
- type inference from literal to type
- type convertability

This makes things much easier to reason about.

There is more work to be done (and TODOs with details are in the code),
but this is a good step forward.

Test: make -j 12 tools HOST_USE_ASAN=true && ./build-x64/host_tests/fidl-compiler-test
Change-Id: Ife8d154f6185b78da8976ed5336192e97ce9f196