[fit] Implement fit::optional and fit::nullable.
fit::optional implements a minimal subset of std::optional.
We intend to expand it to cover the full API of std::optional
in a future patch and make it an alias for std::optional when
compiling for C++ 17.  For now, this subset covers the behavior
needed by other code in libfit.

fit::nullable is similar to fit::optional but it takes advantage
of the null state of the underlying value (if it has one).
This type is especially useful right now for wrapping closures.

Added some new traits to test for nullability and null values.
Deleted previously used imprecise mechanisms.

Simplified implementation of fit::defer using fit::nullable.
This made it possible to flesh out the interface to cover the
remainder of nullptr operations a client might expect given
that the type could be constructed from null.

Added missing checks for self-assignment, self-move, and self-swap.
Also fixed argument-dependent lookup for all swap() functions.

See README.md for more details.

Test: fit-test
Change-Id: I9035741f032b8dbf419462c295d8ca12f62e0a9f