[fit] Move fit::internal::function to fit::function_impl.
Previously fit::function was an alias for fit::internal::function
so this meant that clients were indirectly accessing a type inside
of the fit::internal namespace which doesn't make sense.
So this type has been moved and renamed to fit::function_impl.

As a side-effect, this means that the associated free functions
for swap and comparison can also be moved to the fit namespace.
Argument-dependent lookup found them just fine but again it was
odd for clients to be indirectly reaching into the internal namespace.

Marked non-extensible types final for good measure.

Test: fit-test
Change-Id: Ic90597c92a5ff5a2844ce800b035ebbce2617358