[services] Missing services equivalent to empty.
This change relaxes the restriction that either 'services' xor
'deprecated-all-services' must be present in cmx. Instead, we allow both
to be missing, with the semantics that no 'services' is equivalent to
'services = []'.

This change was deferred to avoid possible races with the transition to
'deprecated-all-services', but now that transition is complete.

TESTED=sandbox_unittest

Change-Id: I98416490e90551a1ccaac2537177689e76762dd7