[fidl][go] Start Union Tags at 1 instead of 0
This doesn't affect the wire format, only the representation of FIDL
unions and only in Go.  It will help detect two kinds of errors:

1. Uninitialized unions were previously of the first member, now they
are errors when marshalling:

union Union1 {
    int64 a;
    int64 b;
};

u1 := Union1{} // This was previously of type "a" with value 0, now it is an error whenmarshalled.

2. Not using setters properly will be caught as errors when
marshalling:

u1 := Union1{}
u1.a = 12345 // This is wrong, it should have been u1.SetA(12345)

Both new and old encodings are modified and interoperable.

TESTED=fx run-test go_fidl_test

Change-Id: Iade64a6e45ffc98865cadd4b4f2ba14013340ce5