[fidl] Redo defaults checking in const declarations
(Revert of revert. Second time around.)

Previously, there were many cases which were not covered, as discussed
in FTP-003 on Defaults. For instance, it was possible to have

    const string s = 5;

Decomposing const declaration type checking into

- can the type be a const at all? e.g. arrays cannot
- type inference from literal to type
- type convertability

This makes things much easier to reason about.

There is more work to be done (and TODOs with details are in the code),
but this is a good step forward.

This change also disallows const declarations for enums

    enum MyEnum : int32 { ... };
    const MyEnum SOME_OTHER_CONSTANT = ...;

This is to reflect that enum constants must be declared directly within
the enum declaration itself.

FIDL-299 #comment

Test: make -j 12 tools HOST_USE_ASAN=true && ./build-x64/host_tests/fidl-compiler-test
Change-Id: I70dd8eb79dd2b535b9e8f44330c18f9bc427c001