[bti] Don't hold bti lock while committing/pinning pages in sys_bti_pin
Don't hold the bti lock while committing and pinning pages in
sys_bti_pin.  In order to facilitate doing this cleanly, create a new
abstraction, |PinnedVmObject|, which is an RAII wrapper around a pinned
range in a vmo.

In the case of resizing a WoD up on ToT, this change yielded a speedup of:

1.24x in latency (mean "bti pin" wall time)
  1245.06 microseconds -> 1000.62 microseconds

4.30x in throughput
  34.33 bytes/microsecond -> 147.66 bytes/microsecond

N=3 resizes on each side.

The throughput increase is so dramatic because previously, requests for small
sizes could get blocked by requests for large sizes.  For example, "bti pin"s
of size 4kb previously had a range of ~[0ms, 6.5ms], and now have a range of
~[0ms, 0.7ms].

MA-493

Test: CQ, boot works, graphics work
Change-Id: I4182bda7e4de98b71c8d191dba31bae2977a9bb0