[rust] Compile debug rust with symbols.
Moves the symbol_level buidl arg to a new file, compiler.gni so it can
be shared across multiple build files.

Hooks up the rust compiler configuration to use this flag, so that the
symbol level is set consistently across languages. Previuosly our rust
code was compiled with no symbols.

Adds a --symbol-level flag to the relevant scripts that build rust. Our
0/1/2 symbol_level values conveniently map to the values accepted by the
"debuginfo" flag of the rust tools.

TEST=it builds, manually checked rust symbols exist in a debug build

Change-Id: I8d60923a45b2729deb5a4a67466e257ddb7cf3f4