[wlan][ath10k] Enable software beaconing
Turns out that QCA9880 does not have HW beacon offload. Instead, the
firmware generates "HOST_SWBA" events when it's time to send a beacon.
The original implementation queries the upper layer to retrieve a beacon
on every such event; we instead use the beacon template that we saved in
the enable_beaconing wlanmac callback.

Note that we don't update TIM yet (and other fields that can change).

NET-1618

TEST=Manually verified that we send beacons in AP mode.

Change-Id: I78522d1f651f8528c8545775708fdfd028ace780