[mediaplayer] fix BufferSet destructor
BufferSets used in FidlDecoder can hold references to PayloadBuffers
on behalf of the outboard decoder. When PayloadBuffers are recycled,
the internals of the BufferSet that created them are interrogated.
This means that it's important that a BufferSet's PayloadBuffers
are destroyed before the BufferSet is destroyed.

While we have a systemic issue with teardown in mediaplayer, and
there remains an issue here (other parties may be holding references
to a BufferSet's PayloadBuffers when the BufferSet is destroyed),
we can at least get rid of the PayloadBuffer references that the
BufferSet itself is holding.

This CL releases all the references it holds on behalf of the
outboard decoder in its destructor, so its PayloadBuffers are
more likely to have been destroyed before the BufferSet's
destructor makes the recycle callbacks untenable.

TEST: manually tested by transitioning from one source to another

Change-Id: If971f54fdb2031b4340bb4c1acab334ef14ae9b9