[debugger] Hook up the new command line parser.
Replaces the existing flags system with a new struct of command line
options that are decoded.

The main console function is separated into a new console_main file and
the existing zxdb main.cc file just redirects to this. This isn't that
important, but it seemed strange to me to call a lot of random console
functions from outside the console directory.

Action setup is separated out from the main initialization and flag
parsing. The existing action creation was moved from flags_impl to
actions.cc, and the unit tests moved to actions_unittest.cc

TEST=unit,manual

Change-Id: I3d431acfacbde357d01a7c3884a0152d3d5ff8d4