[debugger] Hook up symbol search paths.
Adds a command-line switch to add symbol search paths and hooks this up
to the build id index. Some enhancements to the command-line processing
code are added to allow multiple switches of the same name.

A summary of indexed symbols is added to the sym-stat command, which no
longer requires a running procss (if given a running process it will
additionally print out that process' symbol information as before). Some
enhancements to the symbol status output are made to highlight cases
where symbols are missing or not found.

TEST=manual

Change-Id: I854f65bd60ba5dc2888b42f7202c23c64564e39c