[fit] Improve fit::function.
Improved readability by moving the implementation of fit::function
to the public header instead of relying on a SYNOPSIS which
obscured some details.

Moved example code to separate files so it's easier to ensure that
it compiles and runs as the library evolves.

Renamed function_traits to callable_traits since, by design, they
are intended to operate on all manner of callable objects, not
just functions.  Moved the traits to traits.h.

Added is_callable<> along with void_t<> which it depends on.

Added tests to ensure fit::functions are only convertible from
types with compatible signatures.  (ZX-2438)

Test: fit-test
Change-Id: I9db3239c97847d054fcf5f9db684389bb0e7aedf