[user_runner] Designate user environment as first to kill in OOM
This configures the environment for the user to be at the head of the
line for killing when the system encounters a critical low memory
situation. This puts the system into a consistent(ish) state in a
highly visible way in this case instead of the current behavior where
we kill a random job in a less visible way and get error reports for
systems in incoherent states.

Test: run 'k oom lowmem', observe user environment being selected

ZX-2731 #done

Change-Id: I3b60f24d9006214050e377472a52fcbea8a9be12