[rust] Make update-rustc-third-party work on Mac again
Made sure to include the buildtools cmake in the path, since it isn’t always
installed on Macs.

Also set up CFLAGS and LDFLAGs to point to a homebrew installed openssl
for the build script of the openssl crate.

Also changed the run of cargo vendor so that it would put build results in
the Fuchsia out directory instead of in the source tree.

Testing: built on Mac
Change-Id: Ic20d9729632d6f778849a8de8cbc0d1672fa5a54