[images] Add GN arg for manifest file to seed /data with
This adds a GN argument to pass a manifest file in to the creation of
data.blk so that builds can configure extra data to appear in the
initial minfs partition. This is useful for seeding data in like ssh
keys (which is how SDK customers configure ssh today), auth
credentials, default time zone, etc.  The default is to create an empty
partition.

Test: manual
Change-Id: I82a7d05a0158ec12ef48752eac12831e80139053