[wlan][ath10k] add peer association context info to firmware (part 1).
ath10k firmware needs to know about the station before it can accept
packets from the station. This CL will enable the driver to accept
the assoc context from MLME to the firmware.

Note that some fields, especially the rates related, are still hard-coded.
See the TODOs inside the code to find them.

BUG: NET-1528
TEST: With fake wlan_assoc_ctx_t data, the Android phone can ping the AP
      with the following configuration on the AP:

        % ifconfig wlan2 add 192.168.1.1/24
        % wlantool2 ap start 0 -c CHAN --ssid SSID

      On Android phone, click the SSID, then ping 192.168.1.1.

Change-Id: I243684b111862df9bf7e085c2fd7c7632ce57416