[story_storage] Add EntityProvider support to StoryStorage
This adds the ability to store entities in the story storage, and an
empty implementation of the EntityProvider interface.

These entities only have one associated type since the framework does
not do any introspection of the data and thus must always return all the
bytes anyway (i.e. it can't provide different views on the data).

The data and type are stored under separate keys to allow for fetching
of the type without the data. To enable this the two writes are done in
an explicit transaction.

TESTED=story_storage_unittest
MI4-1301 #comment

Change-Id: I834801c82b79ccf39fc14a45d9446c42056f655c