[SetUI] Introduction of Mutation API
The Mutation API is a replacement for the previous update mechanism
for SetUI. Previously, the service inferred changes based on
modifications to the settings object. This interaction is prone to
clobberin and cannot pass handles.

This new approach separates out changes into a dedicated API with
two sets of passed data. The first is a configuration (Mutation union)
that allows for capturing persistable data behind the change. The
second is a set of handles that the client can provide for interactive
mutations.

Test: Initial API introduction, no dependent code in level.
Change-Id: Ia06271f9a0fd621ad1448d5eb3012bae85d2dc90