[fs] Pseudo file Rust library.
First version of a Rust library to help provide ad-hoc files via File
interface from io.fidl.  This is similar to the C++ version in

  zircon/system/ulib/fs/include/fs/pseudo-file.h

I am not sure it should be in a separate crate, as I am going to add a
similar implementation for an ad-hoc directory next.

It has been agreed that we do not want separate unbuffered and buffered
version of the pseudo file.  Buffered version cover all the use cases of
the unbuffered version, but having one version unifies the behaviour.

Also, we have agreed that it is better for the buffered version to
behave more like other kind of files, with as less of a special semantic
as possible.  Because of that this version is slightly different from
what the C++ buffered version is currently doing.  There is only one
buffer, and it is not automatically truncated at any point in time.

US-532

Test: All added unit tests pass.

Change-Id: I6e4bb6d7cb8af2e7c3f8b40cbd8966006397abe1