[wayland] Wire up delete_id events.
The wayland protocol specifies that when the server is done using an
object id, it should emit the wl_display::delete_id event to inform the
client it is now safe to reuse that event.

Wiring up sending of wl_display::delete_id is complicated slightly by
our dependency structure. Generated bindings depend on Client, so this
type is in the core library, but to send the delete_id event we would
like to have access to the wayland core bindings to send this event.

We work around this by allowing the bridge to set callback function
that simply posts a `WlDisplayEvent::DeleteId` event.

This also refactors `ObjectMap` to be the private data structure used
by the `Client` and provides appropriate proxy methods in `Client` to
access the `ObjectMap`.

Test: Verified with a pending xdb_shell implementation that sends lots
      of frame callbacks that object id's are being properly recycled.
Change-Id: Id12ad2e82ee17f486021555ea4f97ebabff3ea6a